ops.argmax supports more types.

+++ mindspore/python/mindspore/common/tensor.py
@@ -2765,8 +2765,7 @@ class Tensor(Tensor_):
             >>> print(a.argmax())
             5
         """
-        # P.Argmax only supports float
-        a = self.astype(mstype.float32)
+        a = self
         if axis is None:
             a = a.ravel()
             axis = 0
